feat: :card_file_box: sample data injection for the database.
in this commit we create the data required for the proper functioning of the api, among these we have:  1. the types of identity documents  2. the different genders  3. the different levels of english according to the european framework.  4. 28 different skills where you can find from programming languages to frameworks.  5. 3 sample users, one being an admin and the other two being companies.  6. 6 example campers with different genders and specific skills.